Add new exclusive-to field to hal entries in VINTF
This new tag signals that the hal entry is expected (matrix) and
provided (manifest) through a specific container on a device.
The first example with "vms":
A VM has clients inside of the VM that expect services
to be available to it from somewhere on the host device like a Secure
Execution Environment. These services are not available to anything
else on the host device no matter what sepolicy exists. They are
exclusive to VMs.

Test: atest LibVintfTest vintf_object_test
Bug: 379673733
Change-Id: I7a0146e9453a0c8d21555e1920baaa89e4bf026f