Add checks directly to vintffm for mandatory framework HALs
This check looks at a given frozen Device Compatibility Matrix and
throws an error if that level of framework manifest doesn't contain the
declared interface.
Some of them are marked optional so camera/TV/autotmotive GSIs can be
built.

This has the same behavior (with slightly different error messages) that
was handled by the 'optional="false"' and 'optional="true"' in the
frozen DCMs. We are getting rid of the 'optional' field, so we needed a
replacement.

Test: remove sensorservice from framework manifest && verify error
Test: remove cameraservice from framework manifest && verify no error
Test: atest -p
Bug: 247575800
Change-Id: Ie7bcf74433751c719a4b0a0bd8ca070c6e62ab6a