Remove 'optional' functionality from libvintf
We no longer pay any attention to the 'optional' attribute in HAL
entries in Matrices. All HALs are optional. Any requirements for which
HALs must be installed are handled in VTS (vts_treble_vintf_vendor_test)
and in the VSR.

We do not throw errors if it exists because that will require a lot of
clean up in our tree and all downstream users.

Test: Add deprecated interface to cuttlefish manifest
Test: Add unknown interface instance to cuttlefish manifest
Test: atest -p
Bug: 247575800

Change-Id: I5df993b604cbbe3a882142772180b46a1972f0aa